+++ src/types/api/index.ts
@@ -1,8 +1,9 @@
 export * as JRPC from './jsonrpc';
 
-export * as RPCSPEC08 from '@starknet-io/starknet-types-08';
 export * as RPCSPEC09 from '@starknet-io/starknet-types-09';
-export { PAYMASTER_API } from '@starknet-io/starknet-types-08';
+export * as RPCSPEC010 from '@starknet-io/starknet-types-010';
+
+export { PAYMASTER_API } from '@starknet-io/starknet-types-010';
 
 // Default export
 // alias for "export * from '@starknet-io/starknet-types-09';" which is done within ./rpc.ts